% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2sas.R
\name{csv2sas}
\alias{csv2sas}
\title{convert csv to xpt and sas7bdat (Windows-only)}
\usage{
csv2sas(wd = getwd(), xptPath = getwd(), sas7bdatPath = getwd(),
  output = "csv2sas.sas", openfile = FALSE)
}
\arguments{
\item{wd}{working directory containing csv files}

\item{output}{sas file output}

\item{openfile}{if TRUE, output file will be opened}

\item{xpt}{directory of xpt files}

\item{sas7bdat}{directory of sas7bdat files}
}
\value{
sas file which can be run inside sas to make genuine xpt and sas7bdat files generated by sas software
}
\description{
convert csv to xpt and sas7bdat by  sas script file convert_csv2sas.sas which can be run inside sas to make genuine xpt and sas7bdat files generated by sas software
}
\examples{
\donttest{
  write.csv(Theoph, 'Theoph.csv', row.names = FALSE)
  write.csv(Indometh, 'Indometh.csv', row.names = FALSE)
  csv2sas()  # Means "Current directory"
}
}
